This probably isn't *that* important but it did speed things up a noticeable amount in my local testing, from 4.5 to 1.2s for first call, and from ~0.4 to ~0.15s for subsequent calls. It'll hit the apps service a tiny bit harder since it means three simultaneous calls rather than three serial calls, but the get-workspace one is the hardest hitting call anyway.