@ianmcorvidae 2016-09-08
I'm closing this in favor of https://github.com/cyverse-de/terrain/pull/3